fix PylonDriver: Copy image before returning
The way the `cv::Mat` is constructed, it does not copy the data but
points to the given buffer of `pylon_image_`.  The latter gets
overwritten every time the method is called, though.  So to ensure that
the returned images do not randomly change, explicitly clone the image.

With this fix, the pylon image also does not need to be a class variable
anymore, so move it to the local scope.