record_image_dataset: Do not downsample by default
Replace the `--no-downsample` argument with `--downsample`, that is do
not downsample by default but only when explicitly requested.

The reasoning is that this way the damage is much smaller when
accidentally omitting the argument (full-sized images can still be
downsampled later without loss of information).

Also properly shut down the camera_backend in the end.