Add script `check_camera_sharpness.py`
Shows a live images of a Pylon camera and results of the Canny-based
sharpness estimation.  Parameters can be adjusted via sliders.

This can be used to play around with the parameters as well as quickly
check if the connected camera passes the sharpness test for a given set
of parameters or not.