Add CameraSettings struct and function to load from file
Add struct `CameraSettings` which is supposed to hold all configurable
settings of the cameras (currently frame rate and the path to the pylon
settings file).
Further add functions to get default values and to load from TOML files.

To avoid having to either pass around the path to the TOML file in
higher-level scripts or to assume a hard-coded default path, use an
environment variable TRICAMERA_CONFIG_FILE to specify the path.

Note: This commit only adds the struct but does not yet use it.  It is
the first step in getting rid of the hard-coded frame rate.