Use Settings in TriCameraDriver
Expect a `Settings` instance as argument (optional, by default use the
"load file from env variable"-constructor).
Use it to set the `rate` at which images are fetched.
Also pass it on to PylonDriver instances, so the file is only read once.

This now technically allows to change the frame rate of the
TriCameraDriver.  Note however, that currently the rate of 10Hz is still
assumed in some other parts of the code, so changing it via the
configuration may not yet work properly in all cases.