fix: Open camera when no device_user_id is given & refactor
Move the code for connecting to a Pylon camera out of the constructor of
PylonDriver to a standalone function.  This allows reusing the code.

While doing this, I noticed that the steps to open the camera and start
grabbing images where accidentally done in the `else`-block handling the
case when a DeviceUserID is specified.  I remember that connecting to a
camera without user id did work in the past, probably due to that.
Moving the relevant function calls out of the if/else should fix it.