(check_camera_sharpness) Do not assume camera rate
The previous implementation seemed to assume a hard-coded frame rate of
10 fps.  This wasn't even correct as PylonDriver uses the camera frame
rate which is configured via the Pylon settings file (currently 100fps).

Instead add a parameter "--update-freq" (still defaulting to 10) and use
that to compute the sleep duration of `cv2.waitKey`.
Behaviour should be mostly the same as before (with default values, the
sleep duration just changes from 90 to 100) but no explicit camera frame
rate is assumed anymore.