Add sensor info for PylonDriver
Add a CameraInfo struct which contains the calibration parameters as
well as the frame rate.  Use this as output for `get_sensor_info()`.

This requires to provide the calibration file to the constructor of the
driver.  The old constructor is still kept for backward compatibility.
In this case, the calibration parameter fields are set to zero.
Frame rate and image size are queried from the camera directly, so these
fields are always set.