Check if image size matches with calibration file
Check if the image size provided in the calibration file matches with
what is reported by the camera.  If they don't match, this is likely due
to the wrong calibration file being used, so raise an error.

If initialized without calibration file, the width/height in
camera_info_ should be default-initialized to zero, so skip the check in
that case.