Add tricamera_log_extract
Simple script to still images from a camera observation log.  This
already exists in trifinger_object_tracking but there wasn't a version
for logs without object data.

To avoid code duplication, implement it as a function that takes the log
reader as input.  With this, the corresponding script in
trifinger_object_tracking can simply import that function and pass the
proper LogReader.