Extend capabilities of CameraCalibrationFile to non-array values
The camera calibration files also contain some non-array fields like
camera name and image size.  To be able to load them from the
CameraCalibrationFile class, extend the `__getitem__` method to check
the type of the field and only try to construct an array if it is a
dict.  Other values are simply returned as is.

Add an explicit method `get_array` to allow being more explicit (e.g. as
a safeguard that will fail if the data does not actually have the
expected structure).