Add camera info in HDF5 log file
The script to convert camera logs to HDF5 files now also requires the
camera parameter files and uses them to add the calibration information
to the HDF5 file.
This makes the files more self-contained and more convenient to handle
(no need to keep the separate calibration files around).

Increase the format version to 2 and update the log viewer script
accordingly (it can work with both versions 1 and 2).

To avoid redundancy, add a tuple with camera names and the tricamera log
magic byte to the `__init__.py` of the package.