Remove the option to downsample images in the driver
The downsampling and following un-demosaicing just degrade image quality
and make it less flexible for the user.  Better drop it and let the user
decide if they want to downsample or not.

Note that this will increase memory requirements for the time series
buffers by factor 4, though.  Hopefully that is not an issue given the
amount of memory our robot computers have.