Initialize last_update_time_ in first call of get_observation
Just initializing it in the constructor doesn't work well, as it can
already be quite out-dated at the first call of get_observation().  This
results in the first calls to always get the observation immediately
without sleep, thus ignoring the configured rate.

With the change of this commit, the first call to get_observation()
should return immediately and from then on adhere to `rate`, no matter
how much time has passed between driver construction and start of the
backend loop.