Add TriCameraLogger with option to save as HDF5
Extends the generic SensorLogger (which relies on cereal serialization
for saving the data) by a method to save the data in HDF5 format.

HDF5 is a standardized format, so it can easily be read without
dependency on our code.  Further, it is self-documenting in so far that
data is organized in named groups and datasets (i.e. can be opened and
analysed without having to know the exact structure of the data).

Finally, it allows to read parts of the data without having to load the
whole file into memory first (as opposed to our custom format).