Extend HDF5 logger to include timeseries timestamps
Add them in a new dataset "sensor_data_timestamps".  Keep the name of
the camera timestamps as just "timestamps" for backwards compatibility.

Also for compatibility, add a minor format version.  The idea is to
increase that if only things are added but existing tools can still read
the files without modifications.  The major "format_version" should only
be updated when making changes that breaks existing tools.

The `tricamera_log_to_hdf5` script is extended accordingly as well.